V1.1

+++ src/Cuadrado.java
@@ -1,7 +1,7 @@
 //Función que se encarga de calcular el area del Cuadrado
 public class Cuadrado extends Rectangulo{
 
-    public Cuadrado(int l1){
+    public Cuadrado(double l1){
         super(l1);
         area = l1*l1;
     }
